radchemoscreens/pyproject.toml
[tool.poetry]
name = "ifitm-cancer-assoc"
version = "0.1.0"
description = ""
authors = ["Jordan Taylor <jtaylor81284@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
playwright = "^1.35.0"
bibtexparser = "^1.4.0"
pandas = "^2.0.2"
openai = "^0.27.8"
openpyxl = "^3.1.2"
xlrd = "^2.0.1"
pydantic = "^1.10.9"
debugger = "^1.3"
scikit-image = "^0.21.0"
beautifulsoup4 = "^4.12.2"
lxml = "^4.9.2"
tabulate = "^0.9.0"
python-dotenv = "^1.0.0"
pyqt5 = "^5.15.9"
bio = "^1.5.9"
mypy = "^1.5.1"
sqlalchemy = "^2.0.20"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.23.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
